% Generated by roxygen2 (4.0.1): do not edit by hand
\name{boxcarBlur}
\alias{boxcarBlur}
\title{Multichannel box car blur}
\usage{
boxcarBlur(n, BA)
}
\arguments{
\item{n}{An integer specifying the number of observations in each channel}

\item{BA}{A numeric vector specifying box car widths for each channel.}
}
\value{
A n by m matrix of normalised box car blur.
}
\description{
Create blur of the box car type.
}
\details{
Creates a matrix of dimension n by m which contains a normalised (unit energy in each column) set of box car blur functions. The generated box-car functions are governed by an input vector \code{BA} which specifies the box car width.
}
\examples{
n <- 128
m <- 3
BA <- 1/sqrt(c(89,121,87))

blurMat <- boxcarBlur(n, BA)
matplot(blurMat, type = 'l', main = paste('Set of box car blur functions'))

signal = makeLIDAR(n)
t = (1:n)/n
matplot(t, signal, type = 'l', main = 'LIDAR test signal')
blurredSignal <- blurSignal(signal, blurMat)

matplot(t, blurredSignal, type = 'l', main = 'Box car blurred LIDAR test signals')
}
\seealso{
\code{\link{gammaBlur}}, \code{\link{blurSignal}}
}

