% Generated by roxygen2 (4.0.1): do not edit by hand
\name{directBlur}
\alias{directBlur}
\title{Direct kernel}
\usage{
directBlur(G.dim)
}
\arguments{
\item{G.dim}{A numeric vector with two elements specifying the values of n and m respectively}
}
\description{
Creates appropriately sized blur matrix for the special case when no blurring is apparent.
}
\details{
Function creates a matrix of dimensin n by m which contains appropriate entries for the case when a direct multichannel signal is observed. That is, no blurring operator is apparent. This is the default argument for the blurring matrix to all the multichannel functions in the \code{mwaved} package.
}

