% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gammaBlur}
\alias{gammaBlur}
\title{Multichannel Gamma density blur}
\usage{
gammaBlur(n, shape, scale)
}
\arguments{
\item{n}{An integer specifying the number of observations in each channel}

\item{shape}{A numeric vector specifying the shape parameter for each Gamma density in each channel}

\item{scale}{A numeric vector specifying the scale parameter for each Gamma density in each channel}
}
\value{
A numeric n by m matrix of normalised Gamma blur.
}
\description{
Create blur of the regular smooth type generated from Gamma densities
}
\details{
Cretae a matrix of dimension n by m which contains a normalised (unit energy in each column) set of blur functions of regular smooth type, generated by a Gamma density. These Gamma densities are generated using the \code{\link{dgamma}} base R function and then normalised to have unit energy.
}
\examples{
n <- 128
m <- 3
shape <- seq(from = 0.5, to = 1, length = m)
scale <- rep(0.25,m)

blurMat <- gammaBlur(n, shape, scale)
matplot(blurMat, type = 'l', main = paste('Set of Gamma', m,'Gamma blur densities.'))

signal = makeLIDAR(n)
t = (1:n)/n
matplot(t, signal, type = 'l', main = 'LIDAR test signal')
blurredSignal <- blurSignal(signal, blurMat)

matplot(t, blurredSignal, type = 'l', main = 'Smooth blurred LIDAR test signals')
}
\seealso{
\code{\link{dgamma}} for details of the Gamma density

\code{\link{boxcarBlur}}, \code{\link{blurSignal}}
}

