% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeSignals}
\alias{makeBlocks}
\alias{makeBumps}
\alias{makeCusp}
\alias{makeDoppler}
\alias{makeLIDAR}
\alias{makeSignals}
\title{Generate test signals for simulation}
\usage{
makeLIDAR(n)

makeBumps(n)

makeDoppler(n)

makeCusp(n)

makeBlocks(n)
}
\arguments{
\item{n}{An integer specifying the length of the desired signal.}
}
\value{
A numeric vector of length n giving the desired test signal.
}
\description{
Generates test signals used the standard literature.
}
\examples{
n <- 1024
signal <- makeLIDAR(n)
t = (1:n)/n
plot(t, signal, main = 'LIDAR test signal', type = 'l')
n <- 1024
signal <- makeLIDAR(n)
t = (1:n)/n
plot(t, signal, main = 'LIDAR test signal', type = 'l')
n <- 1024
signal <- makeDoppler(n)
t = (1:n)/n
plot(t, signal, main = 'Doppler test signal', type = 'l')
n <- 1024
signal <- makeCusp(n)
t = (1:n)/n
plot(t, signal, main = 'Cusp test signal', type = 'l')
n <- 1024
signal <- makeBlocks(n)
t = (1:n)/n
plot(t, signal, main = 'Blocks test signal', type = 'l')
}

