% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{mwaved}
\alias{mwaved}
\alias{mwaved-package}
\title{Multichannel wavelet deconvolution with long memory with mwaved.}
\description{
mwaved Computes the Wavelet deconvolution estimate of a common signal present in multiple channels that have possible different levels of blur and additive error. More information about each function can be found in its help documentation.
}
\details{
mwaved using the WaveD wavelet deconvolution paradigm and is based on the waved package. It generalises the approach by allowing a multichannel signal instead of a single channel signal and allows long memory errors within each channel (independent of each channel). The package also uses the external C FFTW library
}

