% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.waveletCoef}
\alias{plot.waveletCoef}
\title{Multi-Resolution Analysis plot of wavelet coefficients}
\usage{
\method{plot}{waveletCoef}(x, ..., lowest = NULL, highest = NULL,
  thickness = 1, coefTrim = NULL, descending = FALSE, sc = 1)
}
\arguments{
\item{lowest}{Specifies the coarsest resolution to display in the Multi-resolution plot}

\item{highest}{Specifies the finest resolution to display in the Multi-resolution plot}
}
\description{
Plots the wavelet coefficient object in the multiresolution analysis
}

