% Generated by roxygen2 (4.0.2): do not edit by hand
\name{resolutionMethod}
\alias{resolutionMethod}
\title{Select appropriate resolution method for blur type}
\usage{
resolutionMethod(blur)
}
\arguments{
\item{blur}{A character string that specifies the behaviour of the blur function \itemize{
\item 'direct': No blur or direct model is used.
\item 'smooth': Blur that has smooth decay in the Fourier domain.
\item 'box.car': Blur that is of box.car type.
}}
}
\description{
Simple function that maps the blur type to the appropriate resolution selection method
}

