% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sigmaSNR}
\alias{sigmaSNR}
\title{Determine noise scale from \bold{S}ignal to \bold{N}oise \bold{R}atios}
\usage{
sigmaSNR(signal, SNR)
}
\arguments{
\item{signal}{Noisefree multichannel input signal}

\item{SNR}{Desired \bold{S}ignal to \bold{N}oise \bold{R}atio for each channel.}
}
\value{
A numeric vector giving the scales (channel level noise standard deviations) to achieve the desired SNR.
}
\description{
Determine noise scale from \bold{S}ignal to \bold{N}oise \bold{R}atios
}
\details{
TBA
}
\examples{
n <- 1024
m <- 3
signal <- makeLIDAR(n)
blur <- gammaBlur(n, c(0.5, 0.75, 1), rep(1, m))
X <- blurSignal(signal, blur)
SNR <- 10*1:3
sigma <- sigmaSNR(X, SNR)
E <- multiNoise(n, sigma)
sigmaEst <- multiSigma(E)
}
\seealso{
\code{\link{multiNoise}} \code{\link{multiSigma}}
}

