% Generated by roxygen2 (4.0.1): do not edit by hand
\name{waveletThresh}
\alias{waveletThresh}
\title{Apply thresholding regime to a set of wavelet coefficients}
\usage{
waveletThresh(beta, thresh = 0, shrinkType = "hard")
}
\arguments{
\item{beta}{A waveletCoef object that includes a set of n wavelet coeffients and integer
j0 specifying the coarsest resolution.}

\item{thresh}{A numeric vector containing the thresholds to be applied to the coefficients
at each resolution.}

\item{shrinkType}{A character string that specifies which thresholding regime to use.
Available choices are the 'hard', 'soft' or 'garrote'.}
}
\description{
Applies a resolution level thresholding technique to a set of wavelet coefficients,
embedded in a wavelet coefficient object.
}
\details{
Applies one of three specified wavelet thresholding regimes to a wavelet
coefficient object (list with a vector of n wavelet coefficients and an integer
specifying the coarse resolution level). If thresh is not specified, no thresholding
is applied. The formulae applied for 'hard', 'soft' or
'garrote' are given by,\itemize{
 \item Hard: \eqn{
   \delta(x) = x 1(|x| > t)
 }
 \item Soft: \eqn{
   \delta(x) = (x - t) 1(x > t) + (x + t) 1(x > -t)
 }
 \item Garrote: \eqn{
   \delta(x) = (x - t^2/x) 1(|x| > t)
 }
}
where 1 represents the indicator function and \emph{t > 0} represents the threshold.
}

