% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{directBlur}
\alias{directBlur}
\title{Direct kernel matrix}
\usage{
directBlur(n, m = 1)
}
\arguments{
\item{n}{Number of observations in each channel}

\item{m}{Number of channels}
}
\description{
Creates appropriately sized blur matrix for the special case when no blurring is apparent.
}
\details{
Function creates a matrix of dimension n by m which contains appropriate entries for the case when a direct multichannel signal is observed. That is, no blurring operator is apparent. This is the default argument for the blurring matrix to all the multichannel functions in the \code{mwaved} package.
}

