\docType{package}
\name{mwaved}
\alias{mwaved}
\alias{mwaved-package}
\title{mwaved - Multichannel wavelet deconvolution with additive long memory.}
\description{
mwaved provides tools that automate the task of conducting
multichannel wavelet deconvolution. More information about
each function can be found in its individual help
documentation.
}
\details{
Stuff
}
\references{
Garrett Grolemund, Hadley Wickham (2011). Dates and Times
Made Easy with lubridate. Journal of Statistical Software,
40(3), 1-25. \url{http://www.jstatsoft.org/v40/i03/}.
}

