% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.waveletCoef}
\alias{plot.waveletCoef}
\title{Multi-Resolution Analysis plot of wavelet coefficients}
\usage{
\method{plot}{waveletCoef}(beta, lowest = 3, highest = NULL,
  thickness = 1, coefTrim = NULL, descending = FALSE, sc = 1, ...)
}
\arguments{
\item{lowest}{}

\item{j0}{The lowest resolution level to include in the visualisation}

\item{highest}{}
}
\description{
Plots the wavelet coefficients in the multiresolution analysis
}

