% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sigmaSNR}
\alias{sigmaSNR}
\title{Determine noise scale levels from a required \bold{S}ignal to \bold{N}oise \bold{R}atios}
\usage{
sigmaSNR(signal, SNR)
}
\arguments{
\item{signal}{Noisefree multichannel input signal}

\item{SNR}{Desired \bold{S}ignal to \bold{N}oise \bold{R}atio for each channel.}
}
\value{
A numeric vector, sigma of length m, giving the scales (channel level noise standard deviations) to achieve the desired SNR.
}
\description{
Determine noise scale levels from a required \bold{S}ignal to \bold{N}oise \bold{R}atios
}
\details{
The elements of sigma is governed by the blurred \bold{S}ignal-to-\bold{N}oise \bold{R}atio (SNR) measured in decibels (dB) where,
\deqn{SNR  = 10 log_{10} (||k*f||^2/sigma^2)}
and k*f is the blurred signal, ||.|| is the norm operator and sigma is the standard deviation of the noise. Roughly speaking, is considered high, medium and low for the cases 10 dB, 20 dB and 30 dB respectively.
}
\examples{
n <- 1024
m <- 3
signal <- makeLIDAR(n)
blur <- gammaBlur(n, c(0.5, 0.75, 1), rep(1, m))
X <- blurSignal(signal, blur)
SNR <- 10*1:3
sigma <- sigmaSNR(X, SNR)
E <- multiNoise(n, sigma)
sigmaEst <- multiSigma(E)
}
\seealso{
\code{\link{multiNoise}} \code{\link{multiSigma}}
}

